Remove the tests/asmgen part of the test suite
The main use of these tests is when porting to a new instruction set
architecture.  At any other time these tests are redundant with
what is tested elsewhere.  That's the reason they are currently not run,
but only compiled.  Even this has maintenance costs that are not warranted.
Let's remove them entirely.