Use SipHash-1-3 for randomized hash tables
SipHash-1-3 does not have collisions that are independent of the seed,
and therefore provides appropriate resistance against "hash flooding"
attacks on hash tables.

MurmurHash 3 has such collisions, but this is not an issue for
non-randomized hash tables.

We therefore keep using MurmurHash 3 for non-randomized hash tables
because it is faster, especially on 32-bit platforms.

To this end, we add a `version` field to the hash table record and
introduce two new version numbers:
- version 3 uses MurmurHash 3 like version 2 but has the `version`
  field set to 3
- version 4 uses SipHash-1-3 and has the `version` field set to 4.