Module Random: revised initialization of PRNG, continued
We use the MD5 hash function (twice) to create the 256 bits of PRNG
state.  The seed (an array of integers) is serialized to a byte
sequence (each seed element occupies 8 bytes in little-endian
representation) then hashed twice using the MD5 hash function and two
different suffixes.

All this is implemented in OCaml, significantly reducing the amount of
C code in runtime/prng.c.