Add check-symbols-name test
This checks that all symbols exported from the runtime C libraries
are correctly named (prefixed with `caml_` modulo a few exceptions).

Previously, the check was done in the Github/Appveyor CI only.  It
makes sense to do it in the Inria CI too, so that more platforms and
configurations are tested.

As written, the test does nothing for the Windows/MSVC ports.  This is
intentional, as the `check-symbols-name` script probably doesn't work
for MSVC.  However, the libraries produced by the Windows/Mingw ports
are checked, this should be enough.