More careful handling of executable permissions for ocamlc-generated files
Create the output file with RW permissions but not X permissions, to avoid
concurrent execution of a partially-written file.

After the file is complete, add X permissions for each R permissions
that were initially present or that were granted by the user's umask.

For example: if the file did not exist initially and the umask is 027,
the file is created with permissions rw-r-----
then updated to permissions rwx-r-x---

Don't change permissions if the file is not a regular file
(e.g. if it's /dev/null and we run as root...)

Fixes: #11302.