Open redirects from parent + don't log errors from child
In the current code, several errors (e.g. cannot open redirection files)
are reported from the child process, after fork() and before execvp().

This is problematic for two reasons:

1- Error logging uses OCaml I/O channels.  It's not  not sure it's safe to use
   the same I/O channel concurrently in the parent and child processes.
   At a minimum, interleaved output is to be expected.

2- When we use posix_spawn (next commits), errors that occur in the
   child process before exec-ing the command cannot be reported in
   details, they just cause a 127 exit code.

Hence, it's good to do as much processing as possible in the parent process,
so that we can safely report detailed error messages.

This PR moves the opening of files (for redirection of stdout, stderr, stdin)
to the parent process.

The actions that remain in the child (setpgrp, dup2) are expected to
not fail in general, hence we remove all detailed error reporting from
the child process, consistently with what posix_spawn does.