Add functions `Unix.bigread` and `Unix.single_bigwrite`
These are variants of `Unix.read` and `Unix.single_write` that operate
over a big array of characters instead of a byte array.

Since big arrays are allocated outside the heap, we don't need to go
through the 65536-byte bounce buffer, and can read or write more than
65536 bytes in one system call.

Also: document the effect of the bounce buffer on the `read`, `write`
and `single_write` functions.

Fixes: #8352
Closes: #10897